refactor(renderer): lift formatBytes to shared utils
